+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [2.0.0] - 2024-01-19
+### Added
+- Kirby 4 support
+    - updated styles
+- block templates support
+
+### Fixed
+- block controller now gets correct data
+
+
 ## [1.0.2] - 2023-08-12
 ### Changed
 - PageBuilderBLock method toHtml now throws errors instead of returning empty string